+++ spigot/src/main/java/me/tofpu/speedbridge/game/leaderboard/LeaderboardAdapter.java
@@ -5,7 +5,6 @@ import com.google.gson.stream.JsonReader;
 import com.google.gson.stream.JsonToken;
 import com.google.gson.stream.JsonWriter;
 import me.tofpu.speedbridge.api.lobby.BoardUser;
-import me.tofpu.speedbridge.lobby.leaderboard.BoardUserImpl;
 
 import java.io.IOException;
 import java.util.ArrayList;
